Add assistant selector dropdown
Removed assistants from the main navigation - these are now in a dropdown on the main chatbot screen. The conversation will abort and reset all settings when the assistant is changed. The assistant is set to the first canned assistant by default so it doesn't break. The assistant dropdown is searchable and I stuck some CSS on there so it'll scroll as the number of assistants increases.